Move shared code out of specific utils source
