+++ src/edge_utils.c
@@ -41,7 +41,6 @@
 #include <string.h>                  // for memcpy, memset, NULL, memcmp
 #include <sys/time.h>                // for timeval
 #include <sys/types.h>               // for time_t, ssize_t, u_int
-#include <sys/stat.h>                // for mkdir
 #include <time.h>                    // for time
 #include <unistd.h>                  // for gethostname, sleep
 #include "auth.h"                    // for generate_private_key
@@ -3224,86 +3223,6 @@ void edge_term (struct n3n_runtime_data * eee) {
 
 
 /* ************************************** */
-
-static int mkdir_p (const char *pathname, int mode, int uid, int gid) {
-    if(access(pathname, R_OK) == 0) {
-        // it already exists (may not be a dir though)
-        return 0;
-    }
-
-    if(errno != ENOENT) {
-        // some other error
-        return -1;
-    }
-
-#ifndef _WIN32
-    if(mkdir(pathname, mode) == -1) {
-        return -1;
-    }
-    if(chown(pathname, uid, gid) == -1) {
-        return -1;
-    }
-
-#else
-    // Some versions of windows appear to have mkdir(), others _mkdir()
-    // using this gives some undefined warnings, but is most compatible
-    if(_mkdir(pathname) == -1) {
-        return -1;
-    }
-#endif
-
-    return 0;
-}
-
-// TODO: why is this in the edge file? it is used in both edge and supernode
-int n3n_config_setup_sessiondir (n2n_edge_conf_t *conf) {
-    if(!conf->sessionname) {
-        traceEvent(TRACE_NORMAL, "cannot setup sessiondir: no sessionname");
-        return -1;
-    }
-
-    // In the future, once we can run tests on the edge without elevated
-    // permissions, this will probably need a way to check the environment
-    // for the basedir for session dirs
-
-
-#ifndef _WIN32
-    char *basedir = "/run/n3n";
-#endif
-#ifdef _WIN32
-    char basedir[1024];
-    basedir[0] = 0;
-
-    char *userprofile = getenv("USERPROFILE");
-    if(!userprofile) {
-        // TODO: surely, there is a better location?
-        userprofile = getenv("TEMP");
-    }
-    if(!userprofile) {
-        return -1;
-    }
-
-    snprintf(basedir, sizeof(basedir), "%s/n3n", userprofile);
-#endif
-
-    if(mkdir_p(basedir, 0755, conf->userid, conf->groupid) == -1) {
-        traceEvent(TRACE_ERROR, "cannot mkdir %s", basedir);
-        return -1;
-    }
-
-    char buf[1024];
-    snprintf(buf, sizeof(buf), "%s/%s", basedir, conf->sessionname);
-    conf->sessiondir = strdup(buf);
-
-    if(mkdir_p(buf, 0755, conf->userid, conf->groupid) == -1) {
-        traceEvent(TRACE_ERROR, "cannot mkdir %s", buf);
-        return -1;
-    }
-
-    traceEvent(TRACE_NORMAL, "sessiondir: %s", conf->sessiondir);
-
-    return 0;
-}
 
 static int edge_init_sockets (struct n3n_runtime_data *eee) {
